Update Solution for #794. Removed unused imports and moved a class to a new location. Refactoring resulted in a crash without exceptions, so I just recreated the class in the right location
